refactor jsapi - logger
