+++ binding.gyp
@@ -45,8 +45,9 @@
                 "src/commands/coredumper/coredumper.cc",
                 "src/hooks/set_hooks.cc",
                 "src/hooks/fatal_error.cc",
-                "src/jsapi/http.cc",
                 "src/jsapi/configure.cc",
+                "src/jsapi/export_logger.cc",
+                "src/jsapi/http.cc",
             ],
             "include_dirs": [
                 'src',
