+++ src/logger.cc
@@ -8,11 +8,9 @@
 #include <fstream>
 
 #include "configure-inl.h"
-#include "environment_data.h"
 #include "platform/platform.h"
 #include "process_data.h"
 #include "util.h"
-#include "uv.h"
 #include "xpf_mutex-inl.h"
 
 namespace xprofiler {
@@ -135,33 +133,6 @@ void Log(const LOG_LEVEL output_level, const char* type, ThreadId thread_id,
     default:
       break;
   }
-}
-
-#define JS_LOG_WITH_LEVEL(level)                                               \
-  if (!info[0]->IsString() || !info[1]->IsString()) {                          \
-    ThrowTypeError(                                                            \
-        New<String>("log type and content must be string!").ToLocalChecked()); \
-    return;                                                                    \
-  }                                                                            \
-  EnvironmentData* env_data = EnvironmentData::GetCurrent(info);               \
-                                                                               \
-  Local<String> component_string = To<String>(info[0]).ToLocalChecked();       \
-  Utf8String component(component_string);                                      \
-  Local<String> log_content_string = To<String>(info[1]).ToLocalChecked();     \
-  Utf8String log_content(log_content_string);                                  \
-  Log(level, *component, env_data->thread_id(), *log_content);
-
-/* js binding logger */
-void JsInfo(const FunctionCallbackInfo<Value>& info) {
-  JS_LOG_WITH_LEVEL(LOG_INFO)
-}
-
-void JsError(const FunctionCallbackInfo<Value>& info) {
-  JS_LOG_WITH_LEVEL(LOG_ERROR)
-}
-
-void JsDebug(const FunctionCallbackInfo<Value>& info) {
-  JS_LOG_WITH_LEVEL(LOG_DEBUG)
 }
 
 };  // namespace xprofiler
