+++ src/xprofiler.cc
@@ -2,10 +2,10 @@
 #include "environment_data.h"
 #include "hooks/set_hooks.h"
 #include "jsapi/include/configure.h"
+#include "jsapi/include/export_logger.h"
 #include "jsapi/include/http.h"
 #include "library/common.h"
 #include "logbypass/log.h"
-#include "logger.h"
 #include "nan.h"
 #include "platform/platform.h"
 #include "process_data.h"
